Refine project structure
