+++ src/extensions/ranking.py
@@ -42,7 +42,7 @@ from util import ConfigView, FORMATTERS
 from util.exceptions import InvalidConfigException
 from util.extbot import filter_roles, is_role_applied, qualified_name
 from util.resources import R
-from .base import BotExtension
+from overlord.extension import BotExtension
 
 log = logging.getLogger('ranking-extension')
